# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: STACK Service Desk AI Solution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/reports/resolution-rate:
    get:
      operationId: getResolutionRate
      tags: [reports]
      summary: Resolution rate by use case
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resolution rate report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolutionRateReport"
  /v1/reports/sla-compliance:
    get:
      operationId: getSLACompliance
      tags: [reports]
      summary: SLA compliance report
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: SLA compliance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SLAComplianceReport"
...
  /v1/reports/ticket-trends:
    get:
      operationId: getTicketTrends
      tags: [reports]
      summary: Ticket trend analytics
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ticket trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketTrendsReport"
...
  /v1/logs/audit:
    get:
      operationId: getAuditLogs
      tags: [logs]
      summary: Get audit logs
      parameters:
        - name: ticket_id
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
  /v1/logs/api-calls:
    get:
      operationId: getAPICallLogs
      tags: [logs]
      summary: Get API call logs
      parameters:
        - name: api_name
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: API call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APICallLog"
  /v1/logs/powershell:
    get:
      operationId: getPowershellLogs
      tags: [logs]
      summary: Get PowerShell execution logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: PowerShell logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PowershellLog"
...
components:
  schemas:
...
    AuditEvent:
      type: object
      required: [log_id, event_type, actor_type, created_at]
      properties:
        log_id:
          type: string
        ticket_id:
          type: ["string", "null"]
        event_type:
          type: string
        actor_id:
          type: ["string", "null"]
        actor_type:
          type: string
        actor_name:
          type: ["string", "null"]
        details:
          type: ["object", "null"]
          additionalProperties: true
        ip_address:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        created_at:
          type: string
...
    ResolutionRateReport:
      type: object
      required: [by_use_case, overall_auto_pct, overall_manual_pct]
      properties:
        by_use_case:
          type: array
          items:
            $ref: "#/components/schemas/ResolutionRateItem"
        overall_auto_pct:
          type: number
        overall_manual_pct:
          type: number
    ResolutionRateItem:
      type: object
      required: [use_case, total, auto_resolved, manual_resolved, auto_pct]
      properties:
        use_case:
          type: string
        total:
          type: integer
        auto_resolved:
          type: integer
        manual_resolved:
          type: integer
        auto_pct:
          type: number
    SLAComplianceReport:
      type: object
      required: [met, breached, at_risk, met_pct, trend]
      properties:
        met:
          type: integer
        breached:
          type: integer
        at_risk:
          type: integer
        met_pct:
          type: number
        trend:
          type: array
          items:
            $ref: "#/components/schemas/SLATrendPoint"
    SLATrendPoint:
      type: object
      required: [date, met_pct]
      properties:
        date:
          type: string
        met_pct:
          type: number
...
    TicketTrendsReport:
      type: object
      required: [trend]
      properties:
        trend:
          type: array
          items:
            $ref: "#/components/schemas/TicketTrendPoint"
    TicketTrendPoint:
      type: object
      required: [date, total]
      properties:
        date:
          type: string
        total:
          type: integer
        by_use_case:
          type: object
          additionalProperties:
            type: integer
...
    AuditLogListResponse:
      type: object
      required: [logs, total]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        total:
          type: integer
    APICallLog:
      type: object
      required: [api_log_id, api_name, method, called_at]
      properties:
        api_log_id:
          type: string
        ticket_id:
          type: ["string", "null"]
        api_name:
          type: string
        endpoint:
          type: ["string", "null"]
        method:
          type: string
        response_status:
          type: ["integer", "null"]
        duration_ms:
          type: ["integer", "null"]
        called_at:
          type: string
    PowershellLog:
      type: object
      required: [execution_id, ticket_id, script_name, execution_status, executed_at]
      properties:
        execution_id:
          type: string
        ticket_id:
          type: string
        device_name:
          type: ["string", "null"]
        device_ip:
          type: ["string", "null"]
        script_name:
          type: string
        execution_status:
          type: string
        output_log:
          type: ["string", "null"]
        executed_at:
          type: string
        duration_seconds:
          type: ["integer", "null"]
...