# lib/api-zod/src/generated/types/sLATrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 0.1.0
 */

export interface SLATrendPoint {
  date: string;
  met_pct: number;
}