# lib/api-zod/src/generated/types/ticketTrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketTrendPointByUseCase } from './ticketTrendPointByUseCase';

export interface TicketTrendPoint {
  date: string;
  total: number;
  by_use_case?: TicketTrendPointByUseCase;
}